Factor out functions to a library and fix the del bug
